backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
requests
solana
solders
base58
httpx
apscheduler
numpy
aiohttp
websockets
